# Remove allocations from Dns.*