--- a/src/Common/src/Interop/Windows/Kernel32/Interop.FreeLibrary.cs
+++ b/src/Common/src/Interop/Windows/Kernel32/Interop.FreeLibrary.cs
@@ -5,8 +5,6 @@
 using System;
 using System.Runtime.InteropServices;
 
-using Microsoft.Win32.SafeHandles;
-
 internal partial class Interop
 {
     internal partial class Kernel32