--- a/src/Common/src/Interop/Windows/WinSock/AddressInfoHints.cs
+++ b/src/Common/src/Interop/Windows/WinSock/AddressInfoHints.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Runtime.InteropServices;
-
 namespace System.Net.Sockets
 {
     [Flags]