--- a/src/Common/src/Interop/Windows/WinSock/Interop.GetAddrInfoExW.cs
+++ b/src/Common/src/Interop/Windows/WinSock/Interop.GetAddrInfoExW.cs
@@ -13,6 +13,8 @@ internal static partial class Interop
     {
         internal const string GetAddrInfoExCancelFunctionName = "GetAddrInfoExCancel";
 
+        internal const int NS_ALL = 0;
+
         internal unsafe delegate void LPLOOKUPSERVICE_COMPLETION_ROUTINE([In] int dwError, [In] int dwBytes, [In] NativeOverlapped* lpOverlapped);
 
         [DllImport(Libraries.Ws2_32, ExactSpelling = true, CharSet = CharSet.Unicode, SetLastError = true)]
@@ -21,15 +23,30 @@ internal static partial class Interop
             [In] string pServiceName,
             [In] int dwNamespace,
             [In] IntPtr lpNspId,
-            [In] ref AddressInfoEx pHints,
-            [Out] out AddressInfoEx* ppResult,
+            [In] AddressInfoEx* pHints,
+            [Out] AddressInfoEx** ppResult,
             [In] IntPtr timeout,
-            [In] ref NativeOverlapped lpOverlapped,
+            [In] NativeOverlapped* lpOverlapped,
             [In] LPLOOKUPSERVICE_COMPLETION_ROUTINE lpCompletionRoutine,
-            [Out] out IntPtr lpNameHandle
-        );
+            [Out] IntPtr* lpNameHandle);
 
         [DllImport(Libraries.Ws2_32, ExactSpelling = true)]
-        internal static extern unsafe void FreeAddrInfoExW([In] AddressInfoEx* pAddrInfo);
+        internal static extern unsafe void FreeAddrInfoExW(AddressInfoEx* pAddrInfo);
+
+        [StructLayout(LayoutKind.Sequential)]
+        internal unsafe struct AddressInfoEx
+        {
+            internal AddressInfoHints ai_flags;
+            internal AddressFamily ai_family;
+            internal int ai_socktype;
+            internal int ai_protocol;
+            internal IntPtr ai_addrlen;
+            internal IntPtr ai_canonname;    // Ptr to the canonical name - check for NULL
+            internal byte* ai_addr;          // Ptr to the sockaddr structure
+            internal IntPtr ai_blob;         // Unused ptr to blob data about provider
+            internal IntPtr ai_bloblen;
+            internal IntPtr ai_provider;     // Unused ptr to the namespace provider guid
+            internal AddressInfoEx* ai_next; // Next structure in linked list
+        }
     }
 }