--- a/src/Common/src/Interop/Windows/WinSock/Interop.SocketConstructorFlags.cs
+++ b/src/Common/src/Interop/Windows/WinSock/Interop.SocketConstructorFlags.cs
@@ -3,8 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Net.Sockets;
-using System.Runtime.InteropServices;
 
 internal static partial class Interop
 {