--- a/src/System.Net.NameResolution/tests/FunctionalTests/GetHostByAddressTest.cs
+++ b/src/System.Net.NameResolution/tests/FunctionalTests/GetHostByAddressTest.cs
@@ -12,7 +12,6 @@ namespace System.Net.NameResolution.Tests
     public class GetHostByAddressTest
     {
         [Fact]
-
         public void DnsObsoleteGetHostByAddress_BadIPString_Throws()
         {
             Assert.Throws<FormatException>(() => Dns.GetHostByAddress("badIPString"));