--- a/src/System.Net.NameResolution/tests/FunctionalTests/TestSettings.cs
+++ b/src/System.Net.NameResolution/tests/FunctionalTests/TestSettings.cs
@@ -2,11 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Net;
 using System.Net.Sockets;
-using System.Runtime.InteropServices;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace System.Net.NameResolution.Tests