--- a/src/System.Net.NameResolution/tests/PalTests/Fakes/IPAddressFakeExtensions.cs
+++ b/src/System.Net.NameResolution/tests/PalTests/Fakes/IPAddressFakeExtensions.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics;
-
 namespace System.Net
 {
     internal static class IPAddressFakeExtensions