--- a/src/System.Net.NameResolution/tests/UnitTests/Fakes/FakeNameResolutionPal.cs
+++ b/src/System.Net.NameResolution/tests/UnitTests/Fakes/FakeNameResolutionPal.cs
@@ -2,9 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics;
 using System.Net.Sockets;
-using System.Threading;
+using System.Threading.Tasks;
 
 namespace System.Net
 {
@@ -35,7 +34,7 @@ namespace System.Net
             FakesEnsureSocketsAreInitializedCallCount++;
         }
 
-        internal static SocketError TryGetAddrInfo(string hostName, out IPHostEntry ipHostEntry, out int nativeErrorCode)
+        internal static SocketError TryGetAddrInfo(string name, bool justAddresses, out string hostName, out string[] aliases, out IPAddress[] addresses, out int nativeErrorCode)
         {
             throw new NotImplementedException();
         }
@@ -51,7 +50,7 @@ namespace System.Net
             throw new NotImplementedException();
         }
 
-        internal static void GetAddrInfoAsync(DnsResolveAsyncResult asyncResult)
+        internal static Task GetAddrInfoAsync(string hostName, bool justAddresses)
         {
             throw new NotImplementedException();
         }