… src/Common/src/Interop/Windows/WinSock/AddressInfo.cs
… src/Common/src/Interop/Windows/WinSock/AddressInfoEx.cs
… src/Common/src/Interop/Windows/WinSock/Interop.freeaddrinfo.cs
… src/Common/src/Interop/Windows/WinSock/SafeFreeAddrInfo.cs
… src/System.Net.NameResolution/src/System/Net/DNS.cs
… src/System.Net.NameResolution/src/System/Net/Dns.cs
… src/System.Net.NameResolution/src/System/Net/DnsResolveAsyncResult.cs
… src/System.Net.NameResolution/src/System/Net/NameResolutionPal.Unix.cs
… src/System.Net.NameResolution/src/System/Net/NameResolutionPal.Windows.cs
… src/System.Net.NameResolution/src/System/Net/NameResolutionUtilities.cs
… src/System.Net.NameResolution/tests/FunctionalTests/GetHostEntryTest.cs
… src/System.Net.NameResolution/tests/PalTests/NameResolutionPalTests.cs
… src/System.Net.NameResolution/tests/PalTests/System.Net.NameResolution.Pal.Tests.csproj
… src/System.Net.NameResolution/tests/UnitTests/System.Net.NameResolution.Unit.Tests.csproj